深度法线处理和边缘检测

+++ UnityShader/Assets/shader/PostProcess/BriSatCon.shader
@@ -23,7 +23,7 @@
 			
 			struct v2f{
 				float4 clippos : SV_POSITION;
-				half2 uv : TEXCOORD0;
+				float2 uv : TEXCOORD0;
 			};
 
 			v2f vert(appdata_base i){
